Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -77,8 +77,12 @@ router.put(
         imagePath: imagePath
       });
       console.log(post);
-      Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
+      Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if(result.nModified > 0) {
+          res.status(200).json({ message: "Update successful!" });
+        } else {
+          res.status(401).json({ message: "Not authorized!" });
+        }
       });
     }
   );
@@ -108,9 +112,12 @@ router.get("", (req, res, next)=>{
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted!" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(result => {
+      if(result.n > 0) {
+        res.status(200).json({ message: "Post deleted!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
     });
 });
